Move deleteUser to registration module.
